CustomUser/serializers.py
from rest_framework import serializers
from rest_auth.registration.serializers import RegisterSerializer
from rest_auth.serializers import LoginSerializer as RestAuthLoginSerializer

from .models import User


class LoginSerializer(RestAuthLoginSerializer):
    username = None


class UserSerializer(RegisterSerializer):
    username = None

    def get_cleaned_data(self):
        super(UserSerializer, self).get_cleaned_data()

        return {
            'password1': self.validated_data.get('password1', ''),
            'email': self.validated_data.get('email', ''),
        }


class CustomUserDetailsSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = "__all__"
        read_only_fields = ('email',)

